first logger_controller_usb.py abstraction commit
