+++ mat/logger_controller.py
@@ -1,17 +1,12 @@
 # GPLv3 License
 # Copyright (c) 2018 Lowell Instruments, LLC, some rights reserved
 
-from datetime import datetime
-import os
+from abc import ABC, abstractmethod
+import datetime
+import time
 import re
-from serial import (
-    Serial,
-    SerialException,
-)
-from serial.tools.list_ports import grep
 from mat.calibration_factories import calibration_from_string
 from mat.converter import Converter
-from mat.logger_cmd import LoggerCmd
 from mat.logger_info_parser import LoggerInfoParser
 from mat.sensor_parser import SensorParser
 from mat.utils import four_byte_int
@@ -36,6 +31,7 @@ STOP_CMD = 'STP'
 STOP_WITH_STRING_CMD = 'SWS'
 SYNC_TIME_CMD = 'STM'
 TIME_CMD = 'GTM'
+DEL_FILE_CMD = 'DEL'
 
 SIMPLE_CMDS = [
     FIRMWARE_VERSION_CMD,
@@ -50,100 +46,28 @@ SIMPLE_CMDS = [
 ]
 
 
-PORT_PATTERNS = {
-    'posix': r'(ttyACM0)',
-    'nt': r'^COM(\d+)',
-}
-
-
-TIMEOUT = 5
-
-
-def find_port():
-    try:
-        field = list(grep('2047:08[AEae]+'))[0][0]
-    except (TypeError, IndexError):
-        raise RuntimeError("Unable to find port")
-    pattern = PORT_PATTERNS.get(os.name)
-    if not pattern:
-        raise RuntimeError("Unsupported operating system: " + os.name)
-    return re.search(pattern, field).group(1)
-
-
-class LoggerController(object):
+class LoggerController(ABC):
     def __init__(self):
-        self.is_connected = False
-        self.__port = None
-        self.com_port = None
+        super().__init__()
         self.__callback = {}
         self.calibration = None
         self.converter = None
 
-    def open_port(self, com_port=None):
-        try:
-            com_port = com_port or find_port()
-            if com_port:
-                self._open_port(com_port)
-        except SerialException:
-            self.close()
-        return self.is_connected
+    @abstractmethod
+    def open(self):
+        pass
 
-    def _open_port(self, com_port):
-        if isinstance(self.__port, Serial):
-            self.__port.close()
-        if os.name == 'posix':
-            self.__port = Serial('/dev/' + com_port)
-        else:
-            self.__port = Serial('COM' + str(com_port))
-        self.__port.timeout = TIMEOUT
-        self.is_connected = True
-        self.com_port = com_port
+    @abstractmethod
+    def close(self):
+        pass
 
+    @abstractmethod
     def command(self, *args):
-        if not self.is_connected:
-            return None
-        try:
-            return self.find_tag(self.target_tag(args))
-        except SerialException:
-            self.close()
-            self.is_connected = False
-            return None
-
-    def find_tag(self, target):
-        if not target:
-            return
-        while True:
-            cmd = LoggerCmd(self.__port)
-            if cmd.tag == target or cmd.tag == 'ERR':
-                self.callback('rx', cmd.cmd_str())
-                return cmd.result()
+        pass
 
     def callback(self, key, cmd_str):
         if key in self.__callback:
             self.__callback[key](cmd_str)
-
-    def target_tag(self, args):
-        tag = args[0]
-        data = ''
-        if len(args) == 2:
-            data = str(args[1])
-        length = '%02x' % len(data)
-        if tag == 'sleep' or tag == 'RFN':
-            out_str = tag + chr(13)
-        else:
-            out_str = tag + ' ' + length + data + chr(13)
-        self.__port.reset_input_buffer()
-        self.__port.write(out_str.encode('IBM437'))
-        self.callback('tx', out_str[:-1])
-        if tag == RESET_CMD or tag == 'sleep' or tag == 'BSL':
-            return ''
-        return tag
-
-    def close(self):
-        if self.__port:
-            self.__port.close()
-        self.is_connected = False
-        self.com_port = 0
 
     def load_calibration(self):
         read_size = 38
@@ -223,7 +147,7 @@ class LoggerController(object):
         return int(fsz) if fsz else None
 
     def sync_time(self):
-        datetimeObj = datetime.now()
+        datetimeObj = datetime.datetime.now()
         formattedString = datetimeObj.strftime('%Y/%m/%d %H:%M:%S')
         return self.command(SYNC_TIME_CMD, formattedString)
 
@@ -232,6 +156,55 @@ class LoggerController(object):
 
     def __del__(self):
         self.close()
+
+    def get_timestamp(self):
+        """ Return posix timestamp """
+        date_string = self.command(TIME_CMD)
+        epoch = datetime.datetime(1970, 1, 1)  # naive datetime format
+        logger_time = datetime.datetime.strptime(date_string,
+                                                 '%Y/%m/%d %H:%M:%S')
+        return (logger_time-epoch).total_seconds()
+
+    def delete_file(self, name):
+        self.command(DEL_FILE_CMD, name)
+
+    def start_deployment(self):
+        # give time to msp430 to open SD card and create headers
+        answer = self.command(RUN_CMD)
+        time.sleep(2)
+        return answer
+
+    def stop_deployment(self):
+        # give time to msp430 to close SD card
+        answer = self.command(STOP_CMD)
+        time.sleep(2)
+        return answer
+
+    def get_status(self):
+        return self.command(STATUS_CMD)
+
+    def get_time(self):
+        return datetime.datetime.strptime(
+            self.command(TIME_CMD)[6:], '%Y/%m/%d %H:%M:%S')
+
+    def get_serial_number(self):
+        return self.command(SERIAL_NUMBER_CMD)
+
+    def get_firmware_version(self):
+        return self.command(FIRMWARE_VERSION_CMD)
+
+    def check_time(self):
+        pre_time = self.get_time()
+        synced = False
+        if abs(datetime.datetime.now() - pre_time).total_seconds() > 60:
+            synced = True
+            self.sync_time()
+            post_time = self.get_time()
+        if synced:
+            rv = "\n\tTime synced from {} to {}".format(pre_time, post_time)
+        else:
+            rv = "{}".format(pre_time)
+        return rv
 
 
 def _extract_sd_kb(data):
